2 Changes to track Descriptions
Changed the track descriptions both mentioned DevOps so renamed track 2 description to Advanced Scripting and Tools.

Changed the track 3 description to: DevOps, CI\CD and Security in order to shorten it so that it only takes up one line. This is just an aesthetic change so that everything lines up better at the top.